refactor messages tests
